feat: change file structure cli nextjs template
