+++ code/frameworks/nextjs/template/cli/js/Button/Button.stories.js
@@ -1,9 +1,7 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { Button } from './Button';
+import { Button } from '.';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
-const meta: Meta<typeof Button> = {
+export default {
   title: 'Example/Button',
   component: Button,
   tags: ['autodocs'],
@@ -14,31 +12,28 @@ const meta: Meta<typeof Button> = {
   },
 };
 
-export default meta;
-type Story = StoryObj<typeof Button>;
-
 // More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Primary: Story = {
+export const Primary = {
   args: {
     primary: true,
     label: 'Button',
   },
 };
 
-export const Secondary: Story = {
+export const Secondary = {
   args: {
     label: 'Button',
   },
 };
 
-export const Large: Story = {
+export const Large = {
   args: {
     size: 'large',
     label: 'Button',
   },
 };
 
-export const Small: Story = {
+export const Small = {
   args: {
     size: 'small',
     label: 'Button',
