+++ code/frameworks/nextjs/template/cli/js/Header/Header.stories.js
@@ -1,7 +1,6 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { Header } from './Header';
+import { Header } from '.';
 
-const meta: Meta<typeof Header> = {
+export default {
   title: 'Example/Header',
   component: Header,
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
@@ -11,11 +10,7 @@ const meta: Meta<typeof Header> = {
     layout: 'fullscreen',
   },
 };
-
-export default meta;
-type Story = StoryObj<typeof Header>;
-
-export const LoggedIn: Story = {
+export const LoggedIn = {
   args: {
     user: {
       name: 'Jane Doe',
@@ -23,4 +18,6 @@ export const LoggedIn: Story = {
   },
 };
 
-export const LoggedOut: Story = {};
+export const LoggedOut = {
+  args: {},
+};
