+++ code/frameworks/nextjs/template/cli/js/Page/Page.stories.js
@@ -1,9 +1,7 @@
-import type { Meta, StoryObj } from '@storybook/react';
 import { within, userEvent } from '@storybook/testing-library';
+import { Page } from '.';
 
-import { Page } from './Page';
-
-const meta: Meta<typeof Page> = {
+export default {
   title: 'Example/Page',
   component: Page,
   parameters: {
@@ -12,13 +10,10 @@ const meta: Meta<typeof Page> = {
   },
 };
 
-export default meta;
-type Story = StoryObj<typeof Page>;
-
-export const LoggedOut: Story = {};
+export const LoggedOut = {};
 
 // More on interaction testing: https://storybook.js.org/docs/react/writing-tests/interaction-testing
-export const LoggedIn: Story = {
+export const LoggedIn = {
   play: async ({ canvasElement }) => {
     const canvas = within(canvasElement);
     const loginButton = await canvas.getByRole('button', {
