+++ code/frameworks/nextjs/template/cli/ts-3-8/Header/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Button } from './Button';
+import { Button } from '../Button';
 import './header.css';
 
 type User = {
