+++ code/frameworks/nextjs/template/cli/ts-3-8/Page/Page.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { within, userEvent } from '@storybook/testing-library';
 
-import { Page } from './Page';
+import { Page } from '.';
 
 const meta: Meta<typeof Page> = {
   title: 'Example/Page',
