+++ code/frameworks/nextjs/template/cli/ts-3-8/Page/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../Header';
 import './page.css';
 
 type User = {
