+++ code/frameworks/nextjs/template/cli/ts/Header/Header.stories.ts
@@ -1,6 +1,7 @@
-import { Header } from './Header';
+import type { Meta, StoryObj } from '@storybook/react';
+import { Header } from '.';
 
-export default {
+const meta: Meta<typeof Header> = {
   title: 'Example/Header',
   component: Header,
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
@@ -10,7 +11,11 @@ export default {
     layout: 'fullscreen',
   },
 };
-export const LoggedIn = {
+
+export default meta;
+type Story = StoryObj<typeof Header>;
+
+export const LoggedIn: Story = {
   args: {
     user: {
       name: 'Jane Doe',
@@ -18,6 +23,4 @@ export const LoggedIn = {
   },
 };
 
-export const LoggedOut = {
-  args: {},
-};
+export const LoggedOut: Story = {};
